refactor: Move visibility to categories instead
