+++ common/src/main/java/com/wynntils/services/mapdata/providers/builtin/CategoriesProvider.java
@@ -8,8 +8,11 @@ import com.wynntils.services.map.Label;
 import com.wynntils.services.map.type.CombatKind;
 import com.wynntils.services.map.type.ServiceKind;
 import com.wynntils.services.mapdata.attributes.AbstractMapAttributes;
+import com.wynntils.services.mapdata.attributes.impl.FadingMapVisiblity;
+import com.wynntils.services.mapdata.attributes.impl.NeverMapVisibility;
 import com.wynntils.services.mapdata.attributes.type.MapAttributes;
 import com.wynntils.services.mapdata.attributes.type.MapIcon;
+import com.wynntils.services.mapdata.attributes.type.MapVisibility;
 import com.wynntils.services.mapdata.type.MapCategory;
 import com.wynntils.utils.colors.CommonColors;
 import com.wynntils.utils.colors.CustomColor;
@@ -44,6 +47,8 @@ public class CategoriesProvider extends BuiltInProvider {
     }
 
     private static final class WynntilsCategory implements MapCategory {
+        private static final FadingMapVisiblity DEFAULT_VISIBILITY = new FadingMapVisiblity(0, 100, 6);
+
         @Override
         public String getCategoryId() {
             return "wynntils";
@@ -66,11 +71,24 @@ public class CategoriesProvider extends BuiltInProvider {
                 public int getPriority() {
                     return 500;
                 }
+
+                @Override
+                public MapVisibility getIconVisibility() {
+                    return DEFAULT_VISIBILITY;
+                }
+
+                @Override
+                public MapVisibility getLabelVisibility() {
+                    return DEFAULT_VISIBILITY;
+                }
             };
         }
     }
 
     private static final class ServiceCategory implements MapCategory {
+        private static final FadingMapVisiblity FAST_TRAVEL_VISIBLITY = new FadingMapVisiblity(18, 100, 6);
+        private static final FadingMapVisiblity OTHER_VISIBLITY = new FadingMapVisiblity(57, 100, 6);
+
         private final ServiceKind kind;
 
         private ServiceCategory(ServiceKind kind) {
@@ -109,11 +127,28 @@ public class CategoriesProvider extends BuiltInProvider {
                 public CustomColor getLabelColor() {
                     return CommonColors.GREEN;
                 }
+
+                @Override
+                public MapVisibility getIconVisibility() {
+                    if (kind == ServiceKind.FAST_TRAVEL) {
+                        return FAST_TRAVEL_VISIBLITY;
+                    } else {
+                        return OTHER_VISIBLITY;
+                    }
+                }
+
+                @Override
+                public MapVisibility getLabelVisibility() {
+                    return new NeverMapVisibility();
+                }
             };
         }
     }
 
     private static final class CombatCategory implements MapCategory {
+        private static final FadingMapVisiblity CAVES_VISIBILITY = new FadingMapVisiblity(31, 100, 6);
+        private static final FadingMapVisiblity OTHER_VISIBILITY = new FadingMapVisiblity(19, 100, 6);
+
         private final CombatKind kind;
 
         private CombatCategory(CombatKind kind) {
@@ -152,11 +187,29 @@ public class CategoriesProvider extends BuiltInProvider {
                 public CustomColor getLabelColor() {
                     return CommonColors.GREEN;
                 }
+
+                @Override
+                public MapVisibility getIconVisibility() {
+                    if (kind == CombatKind.CAVES) {
+                        return CAVES_VISIBILITY;
+                    } else {
+                        return OTHER_VISIBILITY;
+                    }
+                }
+
+                @Override
+                public MapVisibility getLabelVisibility() {
+                    return new NeverMapVisibility();
+                }
             };
         }
     }
 
     private static final class PlaceCategory implements MapCategory {
+        private static final FadingMapVisiblity PROVINCE_VISIBILITY = new FadingMapVisiblity(0, 32, 3);
+        private static final FadingMapVisiblity CITY_VISIBILITY = new FadingMapVisiblity(0, 74, 3);
+        private static final FadingMapVisiblity PLACE_VISIBILITY = new FadingMapVisiblity(32, 86, 3);
+
         private final Label.LabelLayer layer;
 
         private PlaceCategory(Label.LabelLayer layer) {
@@ -193,6 +246,15 @@ public class CategoriesProvider extends BuiltInProvider {
                 @Override
                 public int getPriority() {
                     return 700;
+                }
+
+                @Override
+                public MapVisibility getLabelVisibility() {
+                    return switch (layer) {
+                        case PROVINCE -> PROVINCE_VISIBILITY;
+                        case CITY -> CITY_VISIBILITY;
+                        case TOWN_OR_PLACE -> PLACE_VISIBILITY;
+                    };
                 }
             };
         }
